Moved functions from inspector to shared
* Some functions in Inspector.js may need to be re-used so moved it to a shared actions file
